refactor : 6주차 코드 리뷰 반영 및 builder 적용
