--- src/main/java/jeje/work/aeatbe/dto/ReviewDTO.java
@@ -1,20 +1,0 @@
-package jeje.work.aeatbe.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class ReviewDTO {
-
-    private Long id;
-    private int rate;
-    private String content;
-    private UserDTO user;
-    private Long productId;
-
-}
